+++ src/components.d.ts
@@ -21,6 +21,7 @@ import { RowCssPartsFn, WcsCellFormatter, WcsGridAllRowSelectedEventDetails, Wcs
 import { HorizontalStepClickEvent, HorizontalStepConfig, HorizontalStepperMode } from "./components/horizontal-stepper/horizontal-stepper-interface";
 import { AutocompleteTypes, InputChangeEventDetail, TextFieldTypes, WcsInputAutocorrect, WcsInputEnterKeyHint, WcsInputInputMode, WcsInputSize, WcsInputState } from "./components/input/input-interface";
 import { MaterialIconFamily, MaterialIconSize } from "./components/mat-icon/mat-icon-interface";
+import { WcsMessageBackground } from "./components/message/message-interface";
 import { ModalSize } from "./components/modal/modal-interface";
 import { WcsNativeSelectSize } from "./components/native-select/native-select-interface";
 import { CssTypes, WcsSize } from "./shared-types";
@@ -50,6 +51,7 @@ export { RowCssPartsFn, WcsCellFormatter, WcsGridAllRowSelectedEventDetails, Wcs
 export { HorizontalStepClickEvent, HorizontalStepConfig, HorizontalStepperMode } from "./components/horizontal-stepper/horizontal-stepper-interface";
 export { AutocompleteTypes, InputChangeEventDetail, TextFieldTypes, WcsInputAutocorrect, WcsInputEnterKeyHint, WcsInputInputMode, WcsInputSize, WcsInputState } from "./components/input/input-interface";
 export { MaterialIconFamily, MaterialIconSize } from "./components/mat-icon/mat-icon-interface";
+export { WcsMessageBackground } from "./components/message/message-interface";
 export { ModalSize } from "./components/modal/modal-interface";
 export { WcsNativeSelectSize } from "./components/native-select/native-select-interface";
 export { CssTypes, WcsSize } from "./shared-types";
@@ -166,7 +168,7 @@ export namespace Components {
     }
     /**
      * Alerts are used to communicate a state or an action that has been performed.
-     * It has to be used conjunction with the `wcs-alert-drawer` component, or you can use it independently by taking care of 
+     * It has to be used conjunction with the `wcs-alert-drawer` component, or you can use it independently by taking care of
      * the alert visibility
      * @cssprop --wcs-alert-background-color - Background color of the alert
      * @cssprop --wcs-alert-icon-background-color - Background color of the icon
@@ -1626,6 +1628,44 @@ export namespace Components {
         "size": MaterialIconSize;
     }
     /**
+     * Messages are used to communicate contextual information to users.
+     * They can display information, success, warning, or error states.
+     * Unlike alerts, messages are static feedback elements and are not automatically dismissed.
+     * @cssprop --wcs-message-background-color-lightest - Lightest background color of the message
+     * @cssprop --wcs-message-background-color-lighter - Lighter background color of the message
+     * @cssprop --wcs-message-title-color - Text color of the title
+     * @cssprop --wcs-message-subtitle-color - Text color of the subtitle
+     * @cssprop --wcs-message-icon-color - Icon color of the message
+     * @cssprop --wcs-message-title-font-weight - Font weight of the title
+     * @cssprop --wcs-message-subtitle-font-weight - Font weight of the subtitle
+     * @cssprop --wcs-message-title-font-size - Font size of the title
+     * @cssprop --wcs-message-subtitle-font-size - Font size of the subtitle
+     * @cssprop --wcs-message-dismiss-button-color - Color of the dismiss button
+     * @cssprop --wcs-message-border-radius - Border radius of the message
+     * @cssprop --wcs-message-border-width - Border width of the message (apply when `border` attribute is set)
+     * @cssprop --wcs-message-border-color - Border color of the message (apply when `border` attribute is set). Based on the intent of the message
+     * @cssprop --wcs-message-padding - Padding of the message
+     * @cssprop --wcs-message-gap - Gap between icon and content
+     */
+    interface WcsMessage {
+        /**
+          * Defines the background appearance of the message.
+         */
+        "background": WcsMessageBackground;
+        /**
+          * Defines the semantic intent of the message. - Non-disruptive messages (`information`, `success`) use `role="status"` - Disruptive messages (`warning`, `error`) use `role="alert"`
+         */
+        "intent": WcsAlertIntent;
+        /**
+          * Defines whether the message should have a border. The border color is determined by the intent of the message.
+         */
+        "showBorder": boolean;
+        /**
+          * Specifies whether the component should display a close button.
+         */
+        "showCloseButton": boolean;
+    }
+    /**
      * The modal component (also named dialog or popup) is an interface element that appears on top of the page content.
      * Use it to show a message, a confirmation dialog, or any other content like forms.
      * ## Accessibility guidelines 💡
@@ -2657,6 +2697,10 @@ export interface WcsInputCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLWcsInputElement;
 }
+export interface WcsMessageCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLWcsMessageElement;
+}
 export interface WcsModalCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLWcsModalElement;
@@ -2795,7 +2839,7 @@ declare global {
     }
     /**
      * Alerts are used to communicate a state or an action that has been performed.
-     * It has to be used conjunction with the `wcs-alert-drawer` component, or you can use it independently by taking care of 
+     * It has to be used conjunction with the `wcs-alert-drawer` component, or you can use it independently by taking care of
      * the alert visibility
      * @cssprop --wcs-alert-background-color - Background color of the alert
      * @cssprop --wcs-alert-icon-background-color - Background color of the icon
@@ -4023,6 +4067,43 @@ declare global {
         prototype: HTMLWcsMatIconElement;
         new (): HTMLWcsMatIconElement;
     };
+    interface HTMLWcsMessageElementEventMap {
+        "wcsMessageDismiss": void;
+    }
+    /**
+     * Messages are used to communicate contextual information to users.
+     * They can display information, success, warning, or error states.
+     * Unlike alerts, messages are static feedback elements and are not automatically dismissed.
+     * @cssprop --wcs-message-background-color-lightest - Lightest background color of the message
+     * @cssprop --wcs-message-background-color-lighter - Lighter background color of the message
+     * @cssprop --wcs-message-title-color - Text color of the title
+     * @cssprop --wcs-message-subtitle-color - Text color of the subtitle
+     * @cssprop --wcs-message-icon-color - Icon color of the message
+     * @cssprop --wcs-message-title-font-weight - Font weight of the title
+     * @cssprop --wcs-message-subtitle-font-weight - Font weight of the subtitle
+     * @cssprop --wcs-message-title-font-size - Font size of the title
+     * @cssprop --wcs-message-subtitle-font-size - Font size of the subtitle
+     * @cssprop --wcs-message-dismiss-button-color - Color of the dismiss button
+     * @cssprop --wcs-message-border-radius - Border radius of the message
+     * @cssprop --wcs-message-border-width - Border width of the message (apply when `border` attribute is set)
+     * @cssprop --wcs-message-border-color - Border color of the message (apply when `border` attribute is set). Based on the intent of the message
+     * @cssprop --wcs-message-padding - Padding of the message
+     * @cssprop --wcs-message-gap - Gap between icon and content
+     */
+    interface HTMLWcsMessageElement extends Components.WcsMessage, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLWcsMessageElementEventMap>(type: K, listener: (this: HTMLWcsMessageElement, ev: WcsMessageCustomEvent<HTMLWcsMessageElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLWcsMessageElementEventMap>(type: K, listener: (this: HTMLWcsMessageElement, ev: WcsMessageCustomEvent<HTMLWcsMessageElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLWcsMessageElement: {
+        prototype: HTMLWcsMessageElement;
+        new (): HTMLWcsMessageElement;
+    };
     interface HTMLWcsModalElementEventMap {
         "wcsDialogClosed": void;
     }
@@ -4789,6 +4870,7 @@ declare global {
         "wcs-list-item-properties": HTMLWcsListItemPropertiesElement;
         "wcs-list-item-property": HTMLWcsListItemPropertyElement;
         "wcs-mat-icon": HTMLWcsMatIconElement;
+        "wcs-message": HTMLWcsMessageElement;
         "wcs-modal": HTMLWcsModalElement;
         "wcs-native-select": HTMLWcsNativeSelectElement;
         "wcs-nav": HTMLWcsNavElement;
@@ -4915,7 +4997,7 @@ declare namespace LocalJSX {
     }
     /**
      * Alerts are used to communicate a state or an action that has been performed.
-     * It has to be used conjunction with the `wcs-alert-drawer` component, or you can use it independently by taking care of 
+     * It has to be used conjunction with the `wcs-alert-drawer` component, or you can use it independently by taking care of
      * the alert visibility
      * @cssprop --wcs-alert-background-color - Background color of the alert
      * @cssprop --wcs-alert-icon-background-color - Background color of the icon
@@ -6405,6 +6487,48 @@ declare namespace LocalJSX {
         "size"?: MaterialIconSize;
     }
     /**
+     * Messages are used to communicate contextual information to users.
+     * They can display information, success, warning, or error states.
+     * Unlike alerts, messages are static feedback elements and are not automatically dismissed.
+     * @cssprop --wcs-message-background-color-lightest - Lightest background color of the message
+     * @cssprop --wcs-message-background-color-lighter - Lighter background color of the message
+     * @cssprop --wcs-message-title-color - Text color of the title
+     * @cssprop --wcs-message-subtitle-color - Text color of the subtitle
+     * @cssprop --wcs-message-icon-color - Icon color of the message
+     * @cssprop --wcs-message-title-font-weight - Font weight of the title
+     * @cssprop --wcs-message-subtitle-font-weight - Font weight of the subtitle
+     * @cssprop --wcs-message-title-font-size - Font size of the title
+     * @cssprop --wcs-message-subtitle-font-size - Font size of the subtitle
+     * @cssprop --wcs-message-dismiss-button-color - Color of the dismiss button
+     * @cssprop --wcs-message-border-radius - Border radius of the message
+     * @cssprop --wcs-message-border-width - Border width of the message (apply when `border` attribute is set)
+     * @cssprop --wcs-message-border-color - Border color of the message (apply when `border` attribute is set). Based on the intent of the message
+     * @cssprop --wcs-message-padding - Padding of the message
+     * @cssprop --wcs-message-gap - Gap between icon and content
+     */
+    interface WcsMessage {
+        /**
+          * Defines the background appearance of the message.
+         */
+        "background"?: WcsMessageBackground;
+        /**
+          * Defines the semantic intent of the message. - Non-disruptive messages (`information`, `success`) use `role="status"` - Disruptive messages (`warning`, `error`) use `role="alert"`
+         */
+        "intent"?: WcsAlertIntent;
+        /**
+          * Event emitted when the message is dismissed.
+         */
+        "onWcsMessageDismiss"?: (event: WcsMessageCustomEvent<void>) => void;
+        /**
+          * Defines whether the message should have a border. The border color is determined by the intent of the message.
+         */
+        "showBorder"?: boolean;
+        /**
+          * Specifies whether the component should display a close button.
+         */
+        "showCloseButton"?: boolean;
+    }
+    /**
      * The modal component (also named dialog or popup) is an interface element that appears on top of the page content.
      * Use it to show a message, a confirmation dialog, or any other content like forms.
      * ## Accessibility guidelines 💡
@@ -7447,6 +7571,7 @@ declare namespace LocalJSX {
         "wcs-list-item-properties": WcsListItemProperties;
         "wcs-list-item-property": WcsListItemProperty;
         "wcs-mat-icon": WcsMatIcon;
+        "wcs-message": WcsMessage;
         "wcs-modal": WcsModal;
         "wcs-native-select": WcsNativeSelect;
         "wcs-nav": WcsNav;
@@ -7534,7 +7659,7 @@ declare module "@stencil/core" {
             "wcs-action-bar": LocalJSX.WcsActionBar & JSXBase.HTMLAttributes<HTMLWcsActionBarElement>;
             /**
              * Alerts are used to communicate a state or an action that has been performed.
-             * It has to be used conjunction with the `wcs-alert-drawer` component, or you can use it independently by taking care of 
+             * It has to be used conjunction with the `wcs-alert-drawer` component, or you can use it independently by taking care of
              * the alert visibility
              * @cssprop --wcs-alert-background-color - Background color of the alert
              * @cssprop --wcs-alert-icon-background-color - Background color of the icon
@@ -8360,6 +8485,27 @@ declare module "@stencil/core" {
              * A component used to display a [Material Icon](https://fonts.google.com/icons?icon.set=Material+Icons). Can be useful when used in wcs-grid or in a wcs-button.
              */
             "wcs-mat-icon": LocalJSX.WcsMatIcon & JSXBase.HTMLAttributes<HTMLWcsMatIconElement>;
+            /**
+             * Messages are used to communicate contextual information to users.
+             * They can display information, success, warning, or error states.
+             * Unlike alerts, messages are static feedback elements and are not automatically dismissed.
+             * @cssprop --wcs-message-background-color-lightest - Lightest background color of the message
+             * @cssprop --wcs-message-background-color-lighter - Lighter background color of the message
+             * @cssprop --wcs-message-title-color - Text color of the title
+             * @cssprop --wcs-message-subtitle-color - Text color of the subtitle
+             * @cssprop --wcs-message-icon-color - Icon color of the message
+             * @cssprop --wcs-message-title-font-weight - Font weight of the title
+             * @cssprop --wcs-message-subtitle-font-weight - Font weight of the subtitle
+             * @cssprop --wcs-message-title-font-size - Font size of the title
+             * @cssprop --wcs-message-subtitle-font-size - Font size of the subtitle
+             * @cssprop --wcs-message-dismiss-button-color - Color of the dismiss button
+             * @cssprop --wcs-message-border-radius - Border radius of the message
+             * @cssprop --wcs-message-border-width - Border width of the message (apply when `border` attribute is set)
+             * @cssprop --wcs-message-border-color - Border color of the message (apply when `border` attribute is set). Based on the intent of the message
+             * @cssprop --wcs-message-padding - Padding of the message
+             * @cssprop --wcs-message-gap - Gap between icon and content
+             */
+            "wcs-message": LocalJSX.WcsMessage & JSXBase.HTMLAttributes<HTMLWcsMessageElement>;
             /**
              * The modal component (also named dialog or popup) is an interface element that appears on top of the page content.
              * Use it to show a message, a confirmation dialog, or any other content like forms.