+++ src/components/message/message-interface.ts
@@ -1,0 +1,1 @@
+export type WcsMessageBackground = 'lighter' | 'lightest';